Fix drving issues (fuel consumption, total dist, no fuel, etc)
r5067
2019-10-02

Warning : DO NOT use 'int id = ...' and use 'Integer id = ...'
instead if the id is a resource id or an unit id and it will
call any lookup tables in UnitManager.
OR else, when removing an element from a List, the List's
method remove(int index) would be automatically chosen to use.
List can't tell if the users want to call the method
remove(Object o) or remove(int index).
Alternatively, use casting such as contains((Integer)id)
and remove((Integer)id) to avoid ambiguity.

## DISCUSSION
1. https://github.com/mars-sim/mars-sim/issues/133

## NEW
1. Add estimatedAveFuelConsumption, estimatedTotalCrewWeight,
   cargoCapacity, beginningMass, endMass, and
   getEstimatedAveFuelConsumption() in Vehicle.
   - Replace the use of getIFuelConsumption() with this new
     method in both VehicleMission and TradeUtil.
2. Add getAverageSettlerWeight() in Settlement for computing
   the average weight of a settler in a settlement.
3. Add 'Home' column (signifying the Associated Settlement)
   in Vehicle Tab in Monitor Tool.
   - Add HOME in the constructor and getValueAt() in
     VehicleTableModel.
4. Add status bar at the bottom of the settlement map tool.
   - Add popLabel, xLabel, yLabel in SettlementWindow.
   - Show x & y coordinates of mouse point and population
5. Record properly totalDistance in TravelMission.
   - Create totalDistance.
   - Modify getTotalDistance().
6. Record distanceTravelled in VehicleMission.
   - Create distanceTravelled.
   - Modify getTotalDistanceTravelled().
7. Add new StatusType, namely OUT_OF_FUEL

## CHANGE
1. Modify the fuel consumption - how much fuel is needed for
   a trip.
   - Estimate the total weight before and after a trip.
   - Use getEstimatedFuelConsumption() in VehicleMission for
     estimating resources at the beginning of a trip.
   - Change getIFuelConsumption() calculation in Vehicle.
2. Split lengthy codes in performDisembarkToSettlementPhase()
   to create unloadCargo(), checkPersonStatus() and
   rescueOperation() in RoverMission.
3. Alter electAnOperator() in Airlock to pick the one with
   the best EVA skill/experience to be the operator.
4. Modify the abstract class of Airlock to adopt using a
   person's unique identifier.


## FIX
1. Close the BufferReader brTest in checkXMLFileVersion()
   in SimulationConfig after file reading is done or else
   it the version.txt cannot be deleted.
2. Remove statements of returning 0 in getProbability() in
   EatMeal.
   - A person may be thirsty and not hungry and vice versa
     and need to consider both possibility.
3. Correct checking for EVA suit or Equipment in
   getTradeValue() in TradeUtil.
4. Correct mobilizeVehicle() in OperateVehicle to stop the
   vehicle if running out of methane.
   - Set distanceTraveled to zero or reduce it if not
     having enough methane to propel forward in a given
	 amount of time.
5. Ensure that the unit id is Integer (instead of id) in
   UnitManager and numerous other classes.

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/Unit.java
@@ -143,7 +143,7 @@ public abstract class Unit implements Serializable, UnitIdentifer, Comparable<Un
 		if (this instanceof MarsSurface)
 			return MARS_SURFACE_ID;
 
-		return UNKNOWN_ID;
+		return (Integer) UNKNOWN_ID;
 	}
 	
 	/**
